Menu Main and Pause Management
